Travis: Simplify and clean up gulp invocation
The default 'script' for javascript projects on Travis CI is
'npm test'. This is the preferred entry point, as it will also
be what (future) contributors intuitively try by default after
cloning the repository (npm it; or: npm install && npm test).

The good news is, running 'npm test' on a clone of this repo
already worked by default, because 'npm test' was already set
up in package.json with the 'scripts/test' property to run Gulp.

Note that this change, while not introducing it, has the side-effect
of restoring use of --verbose for the Travis CI build logs.